requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
python-decouple==3.7
pytz==2022.7
sqlparse==0.4.3
